evaluation/python/agent-test-harness/requirements.txt
requests
python-dotenv          
pandas
tqdm
openai>=1.0.0
jinja2>=3.0.0
pytest>=7.0.0
PyJWT
azure-identity==1.17.1